+++ CHANGELOG.md
@@ -12,7 +12,9 @@
     extension
   * [X509.Test.Suite] New module for generating test suites for TLS client
     testing
+  * [X509.Test.Server] New module: simple server for hosting test suites
   * Add `x509.gen.suite` Mix task
+  * Add `x509.test_server` Mix task
 
 ### Fixes